+++ .gitignore
@@ -103,4 +103,3 @@ the-config.cfg
 set_environ.py
 Things to do.txt
 
-set_environ.py
