--- ctip/scheduler-and-mapper/Stop Locations.csv
@@ -1,34 +1,0 @@
-From PDFs,From API,Score,Latitude,Longitude
-14th Street At 1st Avenue,14th Street At 1st Avenue,1,40.731251,-73.982309
-14th Street At 3rd Avenue,,0,40.733153,-73.986982
-14th Street At Avenue A,14th Street At Avenue A,1,40.73047,-73.980474
-14th Street At Avenue B,14th Street At Avenue B,1,40.729474,-73.97813
-14th Street At Irving Place,,0,40.734248,-73.98937
-1st Avenue At 14th Street,1st Avenue At 14th Street,1,40.731487,-73.982384
-1st Avenue At 17th Street,1st Avenue At 17th Street,1,40.733324,-73.981065
-1st Avenue At 24th Street,,0,40.737374,-73.978077
-1st Avenue At 26th Street,,0,40.738203,-73.977564
-20th Street At Loop Exit,20th Street At Loop Exit,1,40.73408,-73.9779
-3rd Avenue At 11th Street,,0,40.733261,-73.981178
-3rd Avenue At 13th Street,,0,40.732954,-73.987531
-3rd Avenue At 14th Street,,0,40.733222,-73.987282
-3rd Avenue At 17th Street,,0,40.734907,-73.986135
-3rd Avenue At 30th Street,,0,40.742955,-73.980102
-6 MetroTech,,0,40.693934,-73.987116
-715 Broadway,715 Broadway,1,40.729088,-73.993757
-721 Broadway,,0,40.729383,-73.993534
-80 Lafayette Street,80 Lafayette Street,1,40.717038,-74.0019
-Avenue C At 14th Street,Avenue C At 14th Street,1,40.728649,-73.975555
-Avenue C At 16th Street,Avenue C At 16th Street,1,40.72982,-73.974708
-Avenue C At 18th Street,Avenue C At 18th Street,1,40.73106,-73.974112
-Broadway At Broome Street,,0,40.721657,-74.000022
-Cadman Plaza At Clark Street,,0,40.696808,-73.991254
-Canal At Lafayette Street,,0,40.71858,-74.000803
-Centre Street At Broome Street,,0,40.720887,-73.99757
-Centre Street At Canal Street,,0,40.718184,-73.999762
-Gramercy Green,Gramercy Green,1,40.739303,-73.982835
-Greenwich Hotel,,0,40.731135,-74.008176
-Lafayette Street At East 4th Street,,0,40.728036,-73.992904
-Lexington Avenue At 31st Street,,0,40.744448,-73.981223
-MetroTech Roadway At Bridge Street,,0,40.692977,-73.985407
-NYU Langone Medical Center,NYU Langone Medical Center,1,40.741674,-73.97491
