--- ctip/scheduler-and-mapper/common_walking_static.py
@@ -1,3 +1,0 @@
-#!/usr/bin/env python3
-from common import file_in_this_dir
-WALKING_TIMES_PICKLE = file_in_this_dir("WalkingStatic.pickle")
