--- ctip/static/app.css
@@ -1,34 +1,0 @@
-form {
-	margin-bottom: 18px;
-}
-label {
-	display: inline-block;
-	margin: 2px 0;
-}
-select, input {
-	max-width: 180px;
-}
-input[type=number] {
-	width: 3em;
-	text-align: right;
-}
-.keep-together {
-	display: inline-block;
-}
-.input-table {
-	display: table;
-}
-.input-table > * {
-	display: table-row;
-}
-.input-table > * > * {
-	display: table-cell;
-	margin: 0 2px;
-}
-.fieldset-footer {
-	margin: 4px 4px 0;
-	text-align: right;
-}
-#fieldset_where .input-table > * > :first-child {
-	width: 80px;
-}
