--- ctip/templates/departures.html
@@ -1,95 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-	<head>
-		<title>{{ document_title }}</title>
-		<meta charset="UTF-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<link
-			rel="stylesheet"
-			type="text/css"
-			href="https://www.wecsam.com/styles/info-simple.css"
-		>
-		<link
-			rel="stylesheet"
-			type="text/css"
-			href="{{ url_for('static', filename='app.css') }}"
-		>
-	</head>
-	<body>
-		<div>
-			<div class="home-link">
-				<a href="https://www.wecsam.com/">&#9668; Home</a>
-			</div>
-			<h1>
-				New York University<br>
-				Campus Transportation Itinerary Planner
-			</h1>
-			<h2>Departure Lister</h2>
-			<p>
-				Select a bus stop and a time to view departure times from that
-				stop after that time.
-			</p>
-			<p>
-				<a href="/">Click here</a> to go back to the itinerary finder.
-			</p>
-			<form method="get" id="input">
-				<fieldset>
-					<legend>Trip Information</legend>
-					<fieldset id="fieldset_where">
-						<legend>Where</legend>
-						<div class="input-table">
-							<label>
-								<span>Origin</span>
-								<select name="orig" id="select-orig">{%
-									for stop in stops
-								%}
-									<option{%
-										if stop == origin
-									%} selected{%
-										endif
-									%}>{{ stop|e }}</option>{%
-									endfor
-								%}
-								</select>
-							</label>
-						</div>
-						<div id="place-footer" class="fieldset-footer"></div>
-					</fieldset>
-					<fieldset>
-						<legend>When</legend>
-						<span class="keep-together">
-							<select id="select_day" name="day">{%
-								for day, selected in weekdays_checked
-							%}
-								<option{% 
-									if selected
-								%} selected{%
-									endif
-								%}>{{ day }}</option>{%
-								endfor
-							%}
-							</select>
-							<input
-								type="time"
-								name="when"
-								id="time_when"
-								value="{{ when }}"
-							>
-						</span>
-						<span id="use-now"></span>
-					</fieldset>
-					<div class="fieldset-footer">
-						<input type="submit" value="List Departures">
-						or
-						<a href="{{ request.path }}">Reset</a>
-					</div>
-				</fieldset>
-			</form>
-			<div id="output">{{ output_escaped|safe }}</div>
-		</div>
-		<script
-			type="text/javascript"
-			src="{{ url_for('static', filename='app.min.js') }}"
-		></script>
-	</body>
-</html>
