+++ login/templates/email/reset_password.txt
