+++ login/templates/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+  <title>Sam's simple app</title>
+</head>
+
+<body>
+  {% if user.is_authenticated %}
+    <p> Hello {{user.email}} </p>
+  {% else %}
+    <p> Hello Stranger! </p>
+  {%endif%}
+
+</body>
+</html>
