+++ login/templates/reset_password_request.html
