+++ login/templates/save_locations.html
@@ -19,8 +19,6 @@
   {% if error %}
     <p> Error: {{error}} </p>
   {%endif%}
-  <a href= "/view"> View Saved Locations </a>
-  <a href="/">Home</a>
 
 </body>
 </html>
