+++ login/templates/view_locations.html
@@ -10,7 +10,6 @@
 	  {% for place in locations %}
 	    <p> {{place.address}} </p>
 	  {% endfor %}
-	  <a href="/">Home</a>
 
 </body>
 </html>
