+++ requirements.txt
@@ -20,7 +20,6 @@ passlib==1.7.1
 PyJWT==1.6.1
 pymongo==3.6.1
 python-dateutil==2.7.2
-pytz==2018.4
 pywin32-ctypes==0.1.2
 requests==2.18.4
 six==1.11.0
